requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
pyyaml==6.0.1
anthropic==0.18.1